refactor: centralize storage logic
